# services/core/docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  description: |
    REST API для української e-commerce платформи.

    ## Автентифікація
    API використовує Bearer token автентифікацію. Додайте header:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Рейт-ліміти
    - 100 запитів/хвилину для автентифікованих користувачів
    - 30 запитів/хвилину для неавтентифікованих

    ## Формат відповідей
    Всі відповіді повертаються в JSON форматі.

    ## Пагінація
    Використовуйте параметри `page` та `limit` для пагінації.
    Максимальний limit: 100.
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging-api.example.com/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Development
...
paths:
...
  /products/{id}:
    get:
      tags:
        - Products
      summary: Отримати товар
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          description: ID товару
          schema:
            type: string
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Products
      summary: Оновити товар
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Товар оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Products
      summary: Видалити товар
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Товар видалено
        '404':
          $ref: '#/components/responses/NotFound'
  /orders:
    get:
      tags:
        - Orders
      summary: Список замовлень
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

    post:
      tags:
        - Orders
      summary: Створити замовлення
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Замовлення створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Отримати замовлення
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/status:
    patch:
      tags:
        - Orders
      summary: Оновити статус замовлення
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
                tracking_number:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Статус оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /cart:
    get:
      tags:
        - Cart
      summary: Отримати кошик
      operationId: getCart
      parameters:
        - name: cart_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер сторінки
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Кількість елементів на сторінці
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Невірний запит
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Не знайдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Помилка валідації
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
    ValidationError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Product:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category_id:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        price:
          type: number
        compare_at_price:
          type: number
        currency:
          type: string
          default: UAH
        stock:
          type: integer
        available:
          type: integer
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        attributes:
          type: object
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductImage:
      type: object
      properties:
        url:
          type: string
        alt:
          type: string
        position:
          type: integer
...
    UpdateProductRequest:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        category_id:
          type: string
        price:
          type: number
        compare_at_price:
          type: number
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        attributes:
          type: object
        is_active:
          type: boolean
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        parent_id:
          type: string
        image:
          type: string
        product_count:
          type: integer
        is_active:
          type: boolean
    OrderStatus:
      type: string
      enum:
        - new
        - confirmed
        - processing
        - shipped
        - delivered
        - cancelled
        - refunded
    Order:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        payment_status:
          type: string
          enum: [pending, paid, failed, refunded]
        customer:
          $ref: '#/components/schemas/Customer'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          type: number
        discount:
          type: number
        shipping_cost:
          type: number
        total:
          type: number
        currency:
          type: string
        shipping_method:
          type: string
        shipping_address:
          $ref: '#/components/schemas/Address'
        tracking_number:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        product_id:
          type: string
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        total:
          type: number
    CreateOrderRequest:
      type: object
      required:
        - items
        - shipping_method
        - shipping_address
      properties:
        customer_id:
          type: string
        customer:
          $ref: '#/components/schemas/CustomerInput'
        items:
          type: array
          items:
            type: object
            required:
              - product_id
              - quantity
            properties:
              product_id:
                type: string
              quantity:
                type: integer
        shipping_method:
          type: string
        shipping_address:
          $ref: '#/components/schemas/Address'
        payment_method:
          type: string
        notes:
          type: string
        promo_code:
          type: string
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        order_count:
          type: integer
        total_spent:
          type: number
        created_at:
          type: string
          format: date-time
    CustomerInput:
      type: object
      required:
        - name
        - phone
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    Address:
      type: object
      required:
        - city
      properties:
        city:
          type: string
        street:
          type: string
        building:
          type: string
        apartment:
          type: string
        postal_code:
          type: string
        nova_poshta:
          type: object
          properties:
            warehouse_ref:
              type: string
            warehouse_number:
              type: string
    Cart:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        item_count:
          type: integer
        promo_code:
          type: string
    CartItem:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
        price:
          type: number
        total:
          type: number
...